requirements.txt
opencv-python
torch==2.4.1
torchvision==0.19.1
tensorboard
numpy
Pillow
future
tqdm
matplotlib
scipy
scikit-learn
opencv-python
Pylette
mediapipe
imutils
fastapi[all]
uvicorn[standard]
sqlmodel
gunicorn
SQLAlchemy
pydantic